Elfer-Tools/requirements.txt
colorama
whois
requests
speedtest-cli